requirements.txt
python-telegram-bot==20.7
Flask
pandas
openpyxl
